feat: expose guided_query in core and document core/Alliance gate defaults
